fix error

+++ Group1_PoEManagement/PoEManagementWeb/Pages/Accounts/Delete.cshtml.cs
@@ -36,7 +36,7 @@ namespace PoEManagementWeb.Pages.Accounts
                 return NotFound();
             }
 
-            Account = accountRepository.GetAccountByID(id);
+            Account = accountRepository.GetAccountByID((int)id);
 
             if (Account == null)
             {
